% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_annotation_percentages}
\alias{add_annotation_percentages}
\title{AAA}
\usage{
add_annotation_percentages(
  data,
  annotation_file,
  annotation_column_use = "external_gene_name",
  annotation_column_target = "gene_biotype",
  assay = "RNA"
)
}
\description{
AAA
}
\details{
AAA
}
