% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fig_label.R
\name{add_letter}
\alias{add_letter}
\title{Add Letter}
\usage{
add_letter(label = NULL, cex = 2, font = 2, plot = T, ...)
}
\arguments{
\item{label}{A letter to be added. Default: "a".}

\item{cex}{The size of the text to be plotted. Default: 2.}

\item{font}{The font of the text Default: 2.}

\item{plot}{The lable to be ploted}

\item{...}{Other parameters passed to `text`}
}
\value{
nonsense
}
\description{
Adds text to the upperleft corner
}
\details{
Adds text to the upperleft corner. Usually a letter such as "a", "b", etc.
}
\examples{
# Adds a simple figure label
plot(0)
add_letter("a")

# Adds labels to each figure (automatically)
figure_labels <- letters
par(mfrow=c(2,2))
for(i in 1:4){
  plot(0)
  add_letter(figure_labels) ; figure_labels <- figure_labels[-1]
}
}
