% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_scale_legend}
\alias{add_scale_legend}
\title{AAA}
\usage{
add_scale_legend(
  x = par("usr")[2],
  y = par("usr")[4],
  width = diff(par("usr")[1:2]),
  height = diff(par("usr")[3:4]),
  labels = c("min", "max"),
  pal = colorRampPalette(c("blue", "navy", "grey95", "firebrick", "red"))(99)
)
}
\description{
AAA
}
\details{
AAA
}
