% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_size_legend}
\alias{add_size_legend}
\title{AAA}
\usage{
add_size_legend(
  x = par("usr")[2],
  y = par("usr")[4] - diff(par("usr")[3:4])/3,
  width = (diff(par("usr")[1:2])/40 + diff(par("usr")[1:2])/20)/2,
  height = diff(par("usr")[3:4])/4,
  labels = paste0(seq(0, 100, length.out = 3), "\%"),
  pal = "black",
  pch = 21,
  min.cex = 0.2,
  max.cex = 2.5
)
}
\description{
AAA
}
\details{
AAA
}
