% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{annotate_clusters}
\alias{annotate_clusters}
\title{AAA}
\usage{
annotate_clusters(
  data,
  clustering,
  annotation_list,
  annotation_name = "cell_type",
  sd_cutoff_1 = 1,
  sd_cutoff_2 = 1
)
}
\description{
AAA
}
\details{
AAA
}
