% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anticoag_df}
\alias{anticoag_df}
\title{Anticoagulant medicine prescribing data}
\format{
\subsection{\code{anticoag_df}}{

A data frame with 31,792 rows and 10 columns:
\describe{
\item{date}{data month, first day of month specified}
\item{chemical}{Anticoagulant medicine}
\item{bnf_code}{Medicine British National Formulary code}
\item{ods_code}{Sub-icb location Organisation Data Service code}
\item{gss_code}{Sub-icb location Government Statistical Service code}
\item{name}{Sub-icb location name}
\item{items}{Number of items prescribed}
\item{quantity}{Number of individual units prescribed e.g. tablets, capsules}
\item{actual_cost}{Cost of medicine after discounts have been applied}
\item{total_list_size}{Number of people registered at GP in Sub-ICB location}
}
}
}
\source{
\url{https://openprescribing.net/}
\url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-created&tags=all(NAC_LOC)}
}
\usage{
anticoag_df
}
\description{
Data showing the monthly primary care usage of 5 anticoagulant medicines: apixaban,
dabigatran etexilate, rivaroxaban, edoxaban, and warfarin sodium. Data is
shown as monthly usage for all 106 sub-ICB locations in England between
October 2017 and September 2022.
}
\keyword{datasets}
