% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{barlist}
\alias{barlist}
\title{AAA}
\usage{
barlist(
  data,
  genes,
  clustering = NULL,
  plot_y_axis = T,
  plot_x_axis = T,
  labels = NULL,
  srt = 0,
  orderby = NULL,
  assay = "RNA",
  ylab = "expression",
  font.main = 1,
  cex.main = 1,
  main = "",
  cex.axis = 1,
  col = "default",
  draw_mean_lines = T,
  ...
)
}
\description{
AAA
}
\details{
AAA
}
