% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{draw_violin}
\alias{draw_violin}
\title{AAA}
\usage{
draw_violin(
  x,
  base = 0,
  method = "log",
  plot_points = F,
  points_method = "proportional",
  smooth = 2,
  col = "grey",
  border = "grey",
  at = 1,
  pt.col = "grey",
  pt.cex = 0.5,
  pt.pch = 16,
  bw = 0.45,
  max_points = 200,
  transparency = NULL
)
}
\description{
AAA
}
\details{
AAA
}
