% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_shape_2022}
\alias{england_shape_2022}
\title{Polygon for England}
\format{
\subsection{\code{england_shape_2022}}{

A single SFC_multipolygon shape
}
}
\source{
\url{https://geoportal.statistics.gov.uk}
}
\usage{
england_shape_2022
}
\description{
Single polygon containing the boundary for England, created by merging the polygons
for all sub_ICB locations.
}
\keyword{datasets}
