% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaf_colours}
\alias{gaf_colours}
\title{NICE brand colours}
\format{
\subsection{\code{gaf_colours}}{

A named vector with 9 colours and associated HEX codes
}
}
\source{
\url{https://analysisfunction.civilservice.gov.uk/policy-store/data-visualisation-colours-in-charts/}
}
\usage{
gaf_colours
}
\description{
A named vector of Government Analysis Function (GAF) colours for use in plots
}
\keyword{datasets}
