% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{graph_abstraction}
\alias{graph_abstraction}
\title{AAA}
\usage{
graph_abstraction(
  data,
  red = "umap",
  clustering,
  graph = "SNN",
  cutoff = 0,
  nn = 0
)
}
\description{
AAA
}
\details{
AAA
}
