% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nice_colours_full}
\alias{nice_colours_full}
\title{NICE brand colours}
\format{
\subsection{\code{nice_colours_full}}{

A named vector with 33 brand colours and associated HEX codes
}
}
\source{
\url{https://indepth.nice.org.uk/our-brand/index.html}
}
\usage{
nice_colours_full
}
\description{
A named vector of NICE brand colours for use in plots
}
\keyword{datasets}
