% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{nice_cols}
\alias{nice_cols}
\title{Select NICE brand colours}
\usage{
nice_cols(...)
}
\arguments{
\item{...}{Colour names}
}
\value{
A named vector containing specified colours
}
\description{
\code{nice_cols()} is used to select individual colours from the NICE brand colours.
}
\examples{
library(niceRplots)

nice_cols()
nice_cols("black_25")
nice_cols("bold_teal_100")
nice_cols("bold_teal_100", "deep_blue_100")
}
