% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_gg_style.R
\name{nice_gg_style}
\alias{nice_gg_style}
\title{nice_gg_style}
\usage{
nice_gg_style()
}
\description{
nice_gg_style
}
\examples{
line_chart <- ggplot(line_df, aes(x = year, y = people)) +
geom_line(colour = "#228096") +
nice_gg_style()
}
