% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_gg_theme.R
\name{nice_gg_theme}
\alias{nice_gg_theme}
\title{nice_gg_theme}
\usage{
nice_gg_theme(legend = "none")
}
\description{
nice_gg_theme
}
\examples{
line_chart <- ggplot(line_df, aes(x = year, y = people)) +
geom_line(colour = "#228096") +
nice_gg_style()
}
