% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{nice_pal}
\alias{nice_pal}
\title{Select NICE palettes}
\usage{
nice_pal(palette = "cat", reverse = FALSE)
}
\arguments{
\item{palette}{name of palette to return.
\itemize{
\item \code{"cat"} (default): Categorical palette using NICE brand colours.
\item \code{"seq"}: Sequential palette using NICE brand colours.
\item \code{"div"}: Diverging palette using NICE brand colours.
\item \code{"cat_gaf"}: Categorical palette using Government Analysis Function colours.
\item \code{"seq_gaf"}: Sequential palette using Government Analysis Function colours.
}}

\item{reverse}{option to reverse palette. Default is \code{FALSE}.}
}
\value{
A vector containing hex colours for the selected palette
}
\description{
\code{nice_pal()} is used to select a NICE palette for use in charts
}
\examples{
library(niceRplots)

nice_pal("cat")
nice_pal("div", reverse = TRUE)
nice_pal("seq")
nice_pal("cat_gaf", reverse = TRUE)
nice_pal("seq_gaf")
}
