% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_bars}
\alias{plot_bars}
\title{AAA}
\usage{
plot_bars(
  data,
  gene,
  clustering,
  assay = "RNA",
  col = c((scales::hue_pal())(8), RColorBrewer::brewer.pal(9, "Set1"),
    RColorBrewer::brewer.pal(8, "Set2")),
  ...
)
}
\description{
AAA
}
\details{
AAA
}
