% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_dots}
\alias{plot_dots}
\title{AAA}
\usage{
plot_dots(
  data,
  genes,
  clustering,
  pal = c("grey90", "grey70", "navy"),
  main = "",
  assay = "RNA",
  srt = 0,
  cex.row = 1,
  cex.col = 1,
  show_grid = T,
  min_size = 0.2,
  show_axis = T,
  add_legend = T,
  ...
)
}
\description{
AAA
}
\details{
AAA
}
