% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_dots_difference}
\alias{plot_dots_difference}
\title{AAA}
\usage{
plot_dots_difference(
  data,
  data2,
  genes,
  clustering,
  clustering2,
  pal = c("blue", "navy", "grey95", "firebrick", "red"),
  main = "",
  assay = "RNA",
  srt = 0,
  cex.row = 1,
  cex.col = 1,
  show_grid = T,
  min_size = 0.5,
  show_axis = T,
  ...
)
}
\description{
AAA
}
\details{
AAA
}
