% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_enrich}
\alias{plot_enrich}
\title{AAA}
\usage{
plot_enrich(
  pathway_name,
  gmt,
  stats,
  enrichment_table = NULL,
  frame = F,
  axes = F,
  xlab = "",
  ylab = "",
  main = NULL,
  cex.main = 1,
  font.main = 1,
  ...
)
}
\description{
AAA
}
\details{
AAA
}
