% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_feat}
\alias{plot_feat}
\title{AAA}
\usage{
plot_feat(
  x,
  red = "umap",
  feat = NULL,
  label = NULL,
  assay = "RNA",
  pch = 16,
  bg = NA,
  font.labels = 1,
  cex.labels = 1,
  cex = 0.3,
  dims = c(1, 2),
  mins = 0,
  add_graph = NULL,
  percent_connections = 1,
  nbin = 400,
  n = 10,
  main = NULL,
  col = c("grey90", "grey80", "grey70", "navy", "black"),
  ...
)
}
\description{
AAA
}
\details{
AAA
}
