% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_heat}
\alias{plot_heat}
\title{AAA}
\usage{
plot_heat(
  data,
  genes,
  order_metadata = NULL,
  annot = NULL,
  cut_max = 2,
  row.cex = 1,
  main = "",
  heat_color = c("grey90", colorRampPalette(c("grey80", "navy", "navy"))(90)),
  ...
)
}
\description{
AAA
}
\details{
AAA
}
