% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_meta}
\alias{plot_meta}
\title{AAA}
\usage{
plot_meta(
  x,
  red = "umap",
  feat = NULL,
  pch = 16,
  cex = 0.3,
  label = F,
  dims = c(1, 2),
  font.labels = 1,
  cex.labels = 1,
  col = c((scales::hue_pal())(8), RColorBrewer::brewer.pal(9, "Set1"),
    RColorBrewer::brewer.pal(8, "Set2"), RColorBrewer::brewer.pal(8, "Accent"),
    RColorBrewer::brewer.pal(9, "Pastel1"), RColorBrewer::brewer.pal(8, "Pastel2")),
  add_graph = NULL,
  percent_connections = 1,
  nbin = 400,
  add_lines = F,
  main = NULL,
  ...
)
}
\description{
AAA
}
\details{
AAA
}
