% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_sankey}
\alias{plot_sankey}
\title{AAA}
\usage{
plot_sankey(
  df,
  plot_labels = T,
  plot_weights = T,
  color_by = 1,
  order_1_by = "NULL",
  order_2_by = "NULL",
  xlim = c(0, 1),
  ylim1 = c(0, 1),
  ylim2 = c(0, 1),
  use_w1 = T,
  use_w2 = T,
  gapv = 0.03,
  gap2v = 0.005,
  smoothing = 0.15,
  nbreaks = 100,
  add = F,
  pal = NULL,
  ...
)
}
\description{
AAA
}
\details{
AAA
}
