% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_plots.R
\name{plot_smoothers}
\alias{plot_smoothers}
\title{AAA}
\usage{
plot_smoothers(
  seuratObject,
  SlingshotCurves,
  gene,
  clustering_use,
  plotOrder = T,
  col = pal,
  colorByPseudotime = T,
  spar = 0.9,
  pt.size = 0.3,
  factorScale = 1.3
)
}
\description{
AAA
}
\details{
AAA
}
