% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_spatial_feat}
\alias{plot_spatial_feat}
\title{AAA}
\usage{
plot_spatial_feat(
  x,
  red = "slice1",
  feat = NULL,
  res = "lowres",
  label = NULL,
  assay = "Spatial",
  plot_tissue = T,
  rescale = T,
  transparency = "",
  pch = 16,
  bg = NA,
  font.labels = 1,
  cex.labels = 1,
  cex = 1,
  mins = NULL,
  add_graph = NULL,
  percent_connections = 1,
  nbin = 400,
  n = 10,
  main = NULL,
  maxs = NULL,
  col = c("grey95", "grey70", "navy", "black"),
  ...
)
}
\description{
AAA
}
\details{
AAA
}
