% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_tree}
\alias{plot_tree}
\title{AAA}
\usage{
plot_tree(
  data,
  slingshot_curves,
  gene,
  rotate90 = F,
  assay = "RNA",
  edge.weights = F,
  pal = c("grey90", "grey70", "blue3", "navy"),
  minsize = 0.5,
  sizefactor = 2,
  ...
)
}
\description{
AAA
}
\details{
AAA
}
