% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{save_matrix_to_HDF5}
\alias{save_matrix_to_HDF5}
\title{AAA}
\usage{
save_matrix_to_HDF5(
  matrix,
  file_name,
  feature_type = "Gene Expression",
  genome = "undefined"
)
}
\description{
AAA
}
\details{
AAA
}
