% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{violins}
\alias{violins}
\title{AAA}
\usage{
violins(
  data,
  gene = NULL,
  clustering = NULL,
  plot_points = T,
  plot_y_axis = T,
  transparency = NULL,
  plot_x_axis = T,
  smooth = 0.1,
  method = "log",
  points_method = "proportional",
  col = c((scales::hue_pal())(8), RColorBrewer::brewer.pal(9, "Set1"),
    RColorBrewer::brewer.pal(8, "Set2")),
  pt.col = "grey",
  pt.cex = 0.5,
  pt.pch = 16,
  bw = 0.7,
  max_points = 200,
  assay = "RNA",
  srt = 0,
  ylab = "expression",
  cex.main = 1,
  main = gene,
  cex.axis = 1,
  ...
)
}
\description{
AAA
}
\details{
AAA
}
