% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fig_label.R
\name{add_letter}
\alias{add_letter}
\title{Adds legend}
\usage{
add_letter(label = "", cex = 2, font = 2, plot = T, ...)
}
\arguments{
\item{label}{the lable to be ploted}

\item{cex}{does nothing}

\item{font}{the lable to be ploted}

\item{plot}{the lable to be ploted}

\item{...}{other parameters to not be used}
}
\value{
nonsense
}
\description{
Adds legend dah (1line)
}
\details{
Super niceR letter (multiple lines)
}
\examples{
#Adds figure label
add_letter("a")
}
