% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_gg_theme.R
\name{nice_gg_theme}
\alias{nice_gg_theme}
\title{Apply the NICE theme to a chart}
\usage{
nice_gg_theme(
  base_size = 12,
  x_title = TRUE,
  y_title = TRUE,
  grid_lines = "y",
  axis_lines = "x",
  axis_ticks = "x",
  show_legend = TRUE,
  panel_border = FALSE,
  remove_axes = FALSE
)
}
\arguments{
\item{base_size}{Minimum font size. Default is \code{12}.}

\item{x_title}{Option to remove the x axis title. If \code{TRUE} (default), the x axis
title will be present and can be renamed using the \code{labs()} function. If set to
\code{FALSE}, the x axis title will be removed.}

\item{y_title}{Option to remove the y axis title. If \code{TRUE} (default), the y axis
title will be present and can be renamed using the \code{labs()} function. If set to
\code{FALSE}, the y axis title will be removed.}

\item{grid_lines}{determines which major grid lines are shown:
\itemize{
\item \code{"y"} (default): show y axis grid lines
\item \code{"x"}: show x axis grid lines
\item \code{"x+y"}: show both x and y axis grid lines
\item \code{"none"}: remove grid lines
}}

\item{axis_lines}{determines which axis lines are shown:
\itemize{
\item \code{"x"} (default): show x axis line
\item \code{"y"}: show y axis line
\item \code{"x+y"}: show both x and y axis lines
\item \code{"none"}: remove axis lines
}}

\item{axis_ticks}{determines which axis ticks are shown:
\itemize{
\item \code{"x"} (default): show ticks on x axis
\item \code{"y"}: show ticks on y axis
\item \code{"x+y"}: show ticks on both x and y axis
\item \code{"none"}: remove ticks
}}

\item{show_legend}{Option to remove the chart legend. If \code{TRUE} (default), the legend
will be present above the chart. If set to \code{FALSE}, the legend will be removed.}

\item{panel_border}{Option to add a panel border. If \code{FALSE} (default), no border
will be present. If set to \code{TRUE}, a panel border will be added.}

\item{remove_axes}{Option to remove the axes. If \code{FALSE} (default), chart axes
will be present. If set to \code{TRUE}, the axes will be removed. This will overwrite
previous axis formatting options}
}
\value{
A ggplot2 chart object
}
\description{
\code{nice_gg_theme()}applies a theme to a ggplot2 chart object to make it
consistent with the NICE style. Changes made are listed below:
\itemize{
\item Changes the chart title font to Lora SemiBold
\item Changes the base font to Inter Regular
\item Applies a hierarchy for text size
\item Adjusts the chart background and margins
\item Adjusts the chart colour scheme
\item Provides an option to remove axis titles
\item Provides an option to alter grid lines
\item Provides an option to alter axis lines
\item Provides an option to alter axis ticks
\item Provides an option to remove the legend
\item Provides an option to add a panel border
\item Provides an option to remove the axes
}
}
\examples{
library(dplyr)
library(ggplot2)

iris_bar_df <- iris \%>\%
group_by(Species) \%>\%
summarise(Sepal.Width = mean(Sepal.Width)) \%>\%
ungroup()

iris_bar <- ggplot(iris_bar_df) +
geom_col(aes(x = Species, y = Sepal.Width), fill = "#228096", colour = "#000000") +
geom_hline(yintercept = 0, linewidth = 1, colour = "#333333") +
scale_y_continuous(expand = c(0, 0), limits = c(0, 4)) +
nice_gg_theme(x_title = FALSE) +
labs(y = "Sepal width")
}
