% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sub_icb_shapes_2022}
\alias{sub_icb_shapes_2022}
\title{Polygons for sub-ICB locations}
\format{
A data frame with 106 rows and 9 columns:
\describe{
\item{sicbl22cd}{Sub-ICB location Government Statistical Service code}
\item{sicbl22nm}{Sub-ICB location name}
\item{bng_e}{Eastings}
\item{bng_n}{Northings}
\item{long}{Longtitude}
\item{lat}{Latitude}
\item{shape_length}{Shape length in meters}
\item{shape_area}{Shape area in square meters}
\item{geometry}{Polygon shape}
}
}
\source{
\url{https://geoportal.statistics.gov.uk/datasets/ons::sub-integrated-care-board-locations-july-2022-en-buc/explore?location=52.809615\%2C-2.452657\%2C8.22}
}
\usage{
sub_icb_shapes_2022
}
\description{
Data frame containing the April 2022 polygon boundaries for all 106 sub-ICB locations
in England. Data taken directly from the Open Geography Portal as a .geoJSON file.
}
\keyword{datasets}
